projects/marketing_agency/dev_products/development/cold_email_automation/requirements.txt
# 콜드메일 자동화 의존성 (zero-signup)
requests>=2.31.0
anthropic>=0.40.0       # Claude API (ARC 기존 키)
beautifulsoup4>=4.12.0  # 홈피 이메일 추출
dnspython>=2.6.0        # DNS MX 이메일 검증
python-dotenv>=1.0.0
